requirements.txt
allure-pytest==2.15.0
allure-python-commons==2.15.0
pytest==8.4.2
requests==2.32.3
Faker==37.12.0